[#52] 책빵VO 분리 및 ViewModel 수정
