+++ B.READ/B.READ/Sources/Presentation/Library/ViewModel/LibraryViewModel.swift
@@ -8,25 +8,9 @@
 import Foundation
 import SwiftUI
 
+
 // MARK: - (C)LibraryViewModel
 final class LibraryViewModel: ObservableObject {
-  
-  // MARK: - (S)LibraryRecordVO
-  struct LibraryRecordVO {
-    let recordID: String
-    let isbn: String
-    let name: String
-    var coverImage: Data?
-    var state: ReadState
-    var heartCount: Int
-    var starCount: Int
-    var percent: Int
-    var memoCount: Int
-    var quoteCount: Int
-    var period: (start: Date?, end: Date?)
-    var isFavorite: Bool
-    var createdAt: Date
-  }
   
   // MARK: - State
   struct LibraryViewState {
@@ -40,7 +24,6 @@ final class LibraryViewModel: ObservableObject {
     var displayRecords: [LibraryRecordVO] = []
     var selectedTab: Int = 0
   }
-  
   @Published var state: LibraryViewState = .init()
   
   // MARK: - Internal Variable
@@ -65,7 +48,11 @@ final class LibraryViewModel: ObservableObject {
     case .onAppear:
       Task {
         await loadRecords()
-        loadTabs()
+        await MainActor.run {
+          loadTabs()
+          filterRecords()
+          sortDisplayRecords()
+        }
       }
     case .selectTab:
       return
@@ -83,16 +70,19 @@ private extension LibraryViewModel {
       let infos: [LibraryCellInfo] = try await self.fetchRecordCellInfo()
       self.records = infos.map {
         // TODO: - totalPages -> totalPage 변수명 변경
+        // TODO: - coverImageData -> coverImage 필요(강제 언래핑X)
         LibraryRecordVO(
-          recordID: $0.record.id,
+          id: $0.record.id,
           isbn: $0.record.isbn,
           name: $0.book.name,
+          coverImage: $0.book.coverImage,
           state: $0.record.state,
           heartCount: $0.record.heartCount,
           starCount: $0.record.starCount,
           percent: Int(Double($0.record.currentPage) / Double($0.book.totalPages) * 100),
           memoCount: $0.record.memoIDs.count,
           quoteCount: $0.record.quoteIDs.count,
+          period: ($0.record.period.startDate, $0.record.period.endDate),
           isFavorite: $0.record.isFavorite,
           createdAt: $0.record.createdAt
         )
